+++ cowbird/api/webhooks/views.py
@@ -8,11 +8,21 @@ from cowbird.api import schemas as s
 from cowbird.api.schemas import ValidOperations
 from cowbird.permissions_synchronizer import Permission
 from cowbird.services.service_factory import ServiceFactory
+from cowbird.utils import get_logger
 
 
 def dispatch(svc_fct):
+    logger = get_logger(__name__)
+    exceptions = []
     for svc in ServiceFactory().get_active_services():
-        svc_fct(svc)
+        # Allow every service to be notified even if one of them throw an error
+        try:
+            svc_fct(svc)
+        except Exception as exception:  # noqa
+            exceptions.append(exception)
+            logger.error("Exception raised while handling event for service [%s] : [%r]", svc.name, exception)
+    if exceptions:
+        raise Exception(exceptions)
 
 
 @s.UserWebhookAPI.post(schema=s.UserWebhook_POST_RequestSchema, tags=[s.WebhooksTag],
@@ -30,12 +40,14 @@ def post_user_webhook_view(request):
                     msg_on_fail=s.UserWebhook_POST_BadRequestResponseSchema.description)
     user_name = ar.get_multiformat_body(request, "user_name")
     if event == ValidOperations.CreateOperation.value:
-        callback_url = ar.get_multiformat_body(request, "callback_url")
+        # FIXME: Tried with ax.URL_REGEX, but cannot match what seems valid urls...
+        callback_url = ar.get_multiformat_body(request, "callback_url", pattern=None)
         try:
             dispatch(lambda svc: svc.user_created(user_name=user_name))
         except Exception:  # noqa
             # If something bad happens, set the status as erroneous in Magpie
             requests.get(callback_url)
+            # TODO: return something else than 200
     else:
         dispatch(lambda svc: svc.user_deleted(user_name=user_name))
     return ax.valid_http(HTTPOk, detail=s.UserWebhook_POST_OkResponseSchema.description)