+++ cowbird/services/impl/geoserver.py
@@ -1,22 +1,27 @@
-from cowbird.request_queue import RequestQueue
+from celery import chain, shared_task
+
+from cowbird.request_task import RequestTask
 from cowbird.services.service import Service
+from cowbird.services.service_factory import ServiceFactory
 
 
 class Geoserver(Service):
     """
-    Keep Geoserver internal representation in synch with the platform.
+    Keep Geoserver internal representation in sync with the platform.
     """
 
     def __init__(self, name, url):
         super(Geoserver, self).__init__(name, url)
-        self.req_queue = RequestQueue()
 
     def get_resource_id(self, resource_full_name):
         # type (str) -> str
         raise NotImplementedError
 
     def user_created(self, user_name):
-        raise NotImplementedError
+        # ..todo: Replace this simple stub with real implementation
+        res = chain(create_workspace.s(user_name),
+                    create_datastore.s("default"))
+        res.delay()
 
     def user_deleted(self, user_name):
         raise NotImplementedError
@@ -26,3 +31,42 @@ class Geoserver(Service):
 
     def permission_deleted(self, permission):
         raise NotImplementedError
+
+    def create_workspace(self, name):
+        # type (Geoserver, str) -> int
+        """
+        Create a new Geoserver workspace.
+
+        @param self: Geoserver instance
+        @param name: Workspace name
+        @return: Workspace id
+        """
+        # TODO
+        return 1
+
+    def create_datastore(self, workspace_id, name):
+        # type (Geoserver, int, str) -> int
+        """
+        Create a new Geoserver workspace.
+
+        @param self: Geoserver instance
+        @param workspace_id: Workspace id where the datastore must be created
+        @param name: Datastore name
+        @return: Datastore id
+        """
+        # TODO
+        return 1
+
+
+@shared_task(bind=True, base=RequestTask)
+def create_workspace(self, name):
+    # type (str) -> int
+    # Avoid any actual logic in celery task handler, only task related stuff should be done here
+    return ServiceFactory().get_service("Geoserver").create_workspace(name)
+
+
+@shared_task(bind=True, base=RequestTask)
+def create_datastore(self, workspace_id, name):
+    # type (int, str) -> int
+    # Avoid any actual logic in celery task handler, only task related stuff should be done here
+    return ServiceFactory().get_service("Geoserver").create_datastore(workspace_id, name)