+++ tests/test_permissions_synchronizer.py
@@ -1,4 +1,5 @@
 import contextlib
+import os
 import tempfile
 import unittest
 
@@ -8,7 +9,6 @@ import yaml
 
 from cowbird.permissions_synchronizer import Permission, PermissionSynchronizer
 from cowbird.services.service_factory import ServiceFactory
-from cowbird.utils import SingletonMeta
 from tests import utils
 
 
@@ -20,82 +20,106 @@ class TestSyncPermissions(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        cls.app = utils.get_test_app()
+        service1 = "Geoserver"
+        service2 = "Thredds"
+        cls.test_services = [service1, service2]
+        cls.res_root = {service1: "/api/workspaces/private/",
+                        service2: "/catalog/birdhouse/workspaces/private/"}
+        cls.sync_perm_name = {service1: ["read"],
+                              service2: ["read", "browse"]}
+        mapping_point_1 = "mapping_point_1"
+        cls.mapped_service = {service1: service2,
+                              service2: service1}
+        data = {
+            "services": {
+                "Magpie": {"active": True},
+                service1: {"active": True},
+                service2: {"active": True}
+            },
+            "sync_permissions": {
+                mapping_point_1: {
+                    "services": {
+                        service1: cls.res_root[service1],
+                        service2: cls.res_root[service2]
+                    },
+                    "permissions_mapping": [
+                        {
+                            service1: cls.sync_perm_name[service1],
+                            service2: cls.sync_perm_name[service2]
+                        }
+                    ]
+                }
+            }
+
+        }
+        cls.cfg_file = tempfile.NamedTemporaryFile(mode="w", suffix=".cfg", delete=False)
+        with cls.cfg_file as f:
+            f.write(yaml.safe_dump(data))
+        cls.app = utils.get_test_app(settings={"cowbird.config_path": cls.cfg_file.name})
 
     @classmethod
     def tearDownClass(cls):
-        # Remove the service instances initialized with the special sync cfg for next tests
-        SingletonMeta._instances.clear()  # pylint: disable=W0212
-        super(TestSyncPermissions, cls).tearDownClass()
+        utils.clear_services_instances()
+        os.unlink(cls.cfg_file.name)
 
     def test_sync(self):
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".cfg") as tmp:
-            service1 = "Geoserver"
-            service2 = "Thredds"
-            test_services = [service1, service2]
-            res_root = {service1: "/api/workspaces/private/",
-                        service2: "/catalog/birdhouse/workspaces/private/"}
-            sync_perm_name = {service1: ["read"],
-                              service2: ["read", "browse"]}
-            mapping_point_1 = "mapping_point_1"
-            mapped_service = {service1: service2,
-                              service2: service1}
+        """
+        This test parses the sync config and checks that when a permission is created in the `PermissionSynchronizer`
+        the proper permission is created for every synchronized service.
 
-            data = {
-                "services": {
-                    "Magpie": {"active": True},
-                    service1: {"active": True},
-                    service2: {"active": True}
-                },
-                "sync_permissions": {
-                    mapping_point_1: {
-                        "services": {
-                            service1: res_root[service1],
-                            service2: res_root[service2]
-                        },
-                        "permissions_mapping": [
-                            {
-                                service1: sync_perm_name[service1],
-                                service2: sync_perm_name[service2]
-                            }
-                        ]
-                    }
-                }
-            }
-            tmp.write(yaml.safe_dump(data))
-            tmp.seek(0)  # back to start since file still open (auto-delete if closed)
-            with contextlib.ExitStack() as stack:
-                stack.enter_context(mock.patch("cowbird.services.impl.magpie.Magpie",
-                                               side_effect=utils.MockMagpieService))
-                stack.enter_context(mock.patch("cowbird.services.impl.geoserver.Geoserver",
-                                               side_effect=utils.MockAnyService))
-                stack.enter_context(mock.patch("cowbird.services.impl.thredds.Thredds",
-                                               side_effect=utils.MockAnyService))
-                utils.get_test_app(settings={"cowbird.config_path": tmp.name})
+        The `PermissionSynchronizer` `create_permission` function will first find if the applied permission exists in
+        the config. Then for every configured service it will obtain the equivalent permission for this service and
+        apply it to the mocked Magpie service. The `outbound_perm` dict of the mocked Magpie service is then checked
+        to validate that every permission that should have been created in Magpie exists.
+        """
+        with contextlib.ExitStack() as stack:
+            stack.enter_context(mock.patch("cowbird.services.impl.magpie.Magpie",
+                                           side_effect=utils.MockMagpieService))
+            stack.enter_context(mock.patch("cowbird.services.impl.geoserver.Geoserver",
+                                           side_effect=utils.MockAnyService))
+            stack.enter_context(mock.patch("cowbird.services.impl.thredds.Thredds",
+                                           side_effect=utils.MockAnyService))
 
-                resource_name = "resource1"
-                magpie = ServiceFactory().get_service("Magpie")
+            magpie = ServiceFactory().get_service("Magpie")
 
-                for svc in test_services:
-                    for perm_name in sync_perm_name[svc]:
-                        permission = Permission(
-                            service_name=svc,
-                            resource_id=0,
-                            resource_full_name=res_root[svc] + resource_name,
-                            name=perm_name,
-                            access="string1",
-                            scope="string2",
-                            user="string3")
-                        PermissionSynchronizer(magpie).create_permission(permission)
-                        assert len(magpie.json()["outbound_perms"]) == len(sync_perm_name[mapped_service[svc]])
-                        for idx, mapped_perm_name in enumerate(sync_perm_name[mapped_service[svc]]):
-                            outbound_perm = magpie.json()["outbound_perms"][idx]
-                            assert outbound_perm.service_name == mapped_service[svc]
-                            assert outbound_perm.resource_id == utils.MockAnyService.ResourceId
-                            assert outbound_perm.resource_full_name == res_root[mapped_service[svc]] + resource_name
-                            assert outbound_perm.name == mapped_perm_name
-                            assert outbound_perm.access == permission.access
-                            assert outbound_perm.scope == permission.scope
-                            assert outbound_perm.user == permission.user
-                        PermissionSynchronizer(magpie).delete_permission(permission)
-                        assert len(magpie.json()["outbound_perms"]) == 0
+            resource_name = "resource1"
+            # Loop over every service having a permission that must be synchronized to another one
+            for svc in self.test_services:
+                for perm_name in self.sync_perm_name[svc]:
+                    # Create the permission for this service that would be provided by `Magpie` as a hook
+                    permission = Permission(
+                        service_name=svc,
+                        resource_id="0",
+                        resource_full_name=self.res_root[svc] + resource_name,
+                        name=perm_name,
+                        access="string1",
+                        scope="string2",
+                        user="string3")
+
+                    # Apply this permission to the synchronizer (this is the function that is tested!)
+                    PermissionSynchronizer(magpie).create_permission(permission)
+
+                    # `magpie`, which is mocked, will store every permission request that should have been done to the
+                    # Magpie service in the `outbound_perms` dict.
+                    assert len(magpie.json()["outbound_perms"]) == len(self.sync_perm_name[self.mapped_service[svc]])
+
+                    # Validate that the mocked `magpie` instance has received a permission request for every mapped
+                    # permission
+                    for idx, mapped_perm_name in enumerate(self.sync_perm_name[self.mapped_service[svc]]):
+                        outbound_perm = magpie.json()["outbound_perms"][idx]
+                        assert outbound_perm.service_name == self.mapped_service[svc]
+                        assert outbound_perm.resource_id == utils.MockAnyService.ResourceId
+                        assert outbound_perm.resource_full_name == \
+                            self.res_root[self.mapped_service[svc]] + resource_name
+                        assert outbound_perm.name == mapped_perm_name
+                        assert outbound_perm.access == permission.access
+                        assert outbound_perm.scope == permission.scope
+                        assert outbound_perm.user == permission.user
+
+                    # This is the second function being tested which should make remove permission requests to `magpie`
+                    # for every mapped permission
+                    PermissionSynchronizer(magpie).delete_permission(permission)
+
+                    # Again the mocked `magpie` instance should remove every permission from its `outbound_perms` rather
+                    # than making the remove permission request to the Magpie service.
+                    assert len(magpie.json()["outbound_perms"]) == 0