+++ tests/test_request_queue.py
@@ -1,7 +1,0 @@
-import pytest
-
-
-@pytest.mark.request_queue
-def test_request_queue():
-    # TODO
-    pass